+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## WIP
 
 - Allow to send custom headers via mailer service
+- Remove choices.js dependency
 
 ## 2.10.2 (2024-12-08)