+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Allow to send custom headers via mailer service
 - Remove choices.js dependency
+- Fix kiss-sticky z-index issue
 
 ## 2.10.2 (2024-12-08)