+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## WIP
+## 2.10.3 (2024-12-30)
 
 - Allow to send custom headers via mailer service
 - Remove choices.js dependency